refactor: move settings ownership to page
