--- src/features/settings/index.ts
@@ -1,3 +1,0 @@
-export { ConfigForm, type ConfigFormFields, type ConfigFormProps } from "./ui/components/ConfigForm";
-export { useAccountOperations } from "./model/hooks/useAccountOperations";
-export { useConfigFormState } from "./model/hooks/useConfigFormState";
