+++ src/pages/settings/ui/SettingsSection.tsx
@@ -1,5 +1,5 @@
 /**
- * @file Defines the settings feature's Settings Section presentation component.
+ * @file Defines the Settings page's Settings Section presentation component.
  * The component renders props and reports user intent through callbacks while data access stays
  * outside the view.
  */
